+++ qpsolvers/solvers/mosek_.py
@@ -28,13 +28,74 @@ sparse problems, in particular for linear or conic quadratic programs.
 """
 
 from typing import Optional, Union
+import warnings
 
 import cvxopt.msk
 import mosek
 import numpy as np
 import scipy.sparse as spa
 
-from .cvxopt_ import cvxopt_solve_qp
+from ..problem import Problem
+from ..solution import Solution
+from .cvxopt_ import cvxopt_solve_problem
+
+
+def mosek_solve_problem(
+    problem: Problem,
+    initvals: Optional[np.ndarray] = None,
+    verbose: bool = False,
+    **kwargs,
+) -> Solution:
+    """
+    Solve a Quadratic Program defined as:
+
+    .. math::
+
+        \\begin{split}\\begin{array}{ll}
+        \\mbox{minimize} &
+            \\frac{1}{2} x^T P x + q^T x \\\\
+        \\mbox{subject to}
+            & G x \\leq h                \\\\
+            & A x = b                    \\\\
+            & lb \\leq x \\leq ub
+        \\end{array}\\end{split}
+
+    using the `MOSEK interface from CVXOPT
+    <https://cvxopt.org/userguide/coneprog.html#optional-solvers>`_.
+
+    Parameters
+    ----------
+    P :
+        Symmetric quadratic-cost matrix.
+    q :
+        Quadratic-cost vector.
+    G :
+        Linear inequality constraint matrix.
+    h :
+        Linear inequality constraint vector.
+    A :
+        Linear equality constraint matrix.
+    b :
+        Linear equality constraint vector.
+    lb :
+        Lower bound constraint vector.
+    ub :
+        Upper bound constraint vector.
+    initvals :
+        Warm-start guess vector.
+    verbose :
+        Set to `True` to print out extra information.
+
+    Returns
+    -------
+    :
+        Solution to the QP, if found, otherwise ``None``.
+    """
+    original_options = cvxopt.solvers.options
+    cvxopt.solvers.options["mosek"] = {mosek.iparam.log: 1 if verbose else 0}
+    solution = cvxopt_solve_problem(problem, "mosek", initvals, **kwargs)
+    cvxopt.solvers.options = original_options
+    return solution
 
 
 def mosek_solve_qp(
@@ -95,8 +156,12 @@ def mosek_solve_qp(
     :
         Solution to the QP, if found, otherwise ``None``.
     """
-    original_options = cvxopt.solvers.options
-    cvxopt.solvers.options["mosek"] = {mosek.iparam.log: 1 if verbose else 0}
-    x = cvxopt_solve_qp(P, q, G, h, A, b, lb, ub, "mosek", initvals, **kwargs)
-    cvxopt.solvers.options = original_options
-    return x
+    warnings.warn(
+        "The return type of this function will change "
+        "to qpsolvers.Solution in qpsolvers v3.0",
+        DeprecationWarning,
+        stacklevel=2,
+    )
+    problem = Problem(P, q, G, h, A, b, lb, ub)
+    solution = mosek_solve_problem(problem, initvals, verbose, **kwargs)
+    return solution.x